Deploy each instance in a different thread
This change applies multithreading to the deployment step to speed up
the process. This way 9 threads wil bmake the deployement to the nine
instances in parallel.

This change adds a finally clause to crotical sections where the threads
may encounter an exception. This way the thread will end its operation
and the script doesn't get blocked in an infinite loop.

Signed-off-by: Mark Bekhet <markonsy72@gmail.com>